Methods

This section describes systems methods implemented in RangeEnclosures.jl.

Methods

The enclose function

Utility functions

Enclosure methods

Specifics to interval optimization

Specifics to sum-of-squares optimization

The enclose function

RangeEnclosures.enclose — Function

enclose(f::Function, dom::Interval_or_IntervalBox,
        solver::Symbol=:IntervalArithmetic; [kwargs]...)

Return a range enclosure of a univariate or multivariate function on the given domain.

Input

f – function

dom – hyperrectangular domain, either a unidimensional Interval or a multidimensional IntervalBox

solver – (optional, default: IntervalArithmetic) choose one among the available solvers; see RangeEnclosures.available_solvers for the full list

kwargs – optional keyword arguments passed to the solver; for available options see the documentation of each solver

Output

An interval representing the range enclosure (minimum and maximum) of f over its domain dom.

Examples

julia> using RangeEnclosures

julia> enclose(x -> 1 - x^4 + x^5, 0..1) # use default solver
[0, 2]

julia> enclose(x -> 1 - x^4 + x^5, 0..1, :IntervalArithmetic)
[0, 2]

julia> enclose(x -> 1 - x^4 + x^5, 0..1, :TaylorModels, order=4)
[0.78125, 1.125]

julia> enclose(x -> 1 - x^4 + x^5, 0..1, :TaylorModels, order=10)
[0.8125, 1.09375]

julia> enclose(x -> 1 - x^4 + x^5, 0..1, :IntervalOptimisation)
[0.916034, 1.00213]

You can also try other solvers such as SumOfSquares and AffineArithmetic.

A vector of solvers can be passed in the solver options. Then, the result is obtained by intersecting the range enclosure of each solver. In the previous example,

julia> using RangeEnclosures

julia> enclose(x -> 1 - x^4 + x^5, 0..1, [:TaylorModels, :IntervalArithmetic])
[0.8125, 1.09375]

Utility functions

RangeEnclosures.relative_precision — Function

relative_precision(x::Interval{N}, xref::Interval{N}) where {N}

Return the relative precision of an interval representing a range enclosure given a reference interval.

Input

x – test interval

xref – reference interval

Output

An interval containing the left (resp. right) percentages that describe the relative precision of x with respect to the reference interval xref.

Examples

Suppose that the reference interval is $[-1.2, 4.6]$, and let $[-1.25, 7.45]$ be the overapproximation of the reference interval obtained with some global optimisation method that returns a range enclosure. The relative precision is then:

julia> using RangeEnclosures: relative_precision

julia> xref = Interval(-1.2, 4.6)
[-1.2, 4.6]

julia> x = Interval(-1.25, 7.45)
[-1.25, 7.45001]

julia> relative_precision(x, xref)
[0.862068, 49.138]

Notice how the percentage of relative error of the maximum is big, close to $50%$, while the approximation of the minimum is much smaller, close to $1%$ of the reference value.

Algorithm

This function measures the relative precision of the result in a more informative way than taking the scalar overestimation because it evaluates the precision of the lower and the upper range bounds separately (cf. Eq. (20) in [1]).

[1] Althoff, Matthias, Dmitry Grebenyuk, and Niklas Kochdumper. Implementation of Taylor models in CORA 2018. Proc. of the 5th International Workshop on Applied Verification for Continuous and Hybrid Systems. 2018.

Notes

It is assumed that the test, or paragon interval, is an overapproximation of the reference interval; otherwise the left or right percentages will be negative.

Enclosure methods

RangeEnclosures.enclose_IntervalArithmetic — Function

enclose_IntervalArithmetic(f::Function, dom::Interval_or_IntervalBox)

Compute a range enclosure using interval arithmetic substitution.

Input

f – function

dom – hyperrectangular domain, either a unidimensional Interval or a multidimensional IntervalBox

Output

An interval representing the range enclosure (minimum and maximum) of f over its domain dom.

Algorithm

The result is obtained by substitution of dom on the function f and application of the rules of interval arithmetic.

We refer to the documentation and source code of IntervalArithmetic for details.

RangeEnclosures.enclose_IntervalOptimisation — Function

enclose_IntervalOptimisation(f::Function, dom::Interval_or_IntervalBox;
                             structure=HeapedVector, tol=1e-3)

Compute a range enclosure using interval optimisation.

Input

f – function

dom – hyperrectangular domain, either a unidimensional Interval or a multidimensional IntervalBox

structure – (optional, default: HeapedVector) the way in which vector elements are kept arranged; possible options are HeapedVector and SortedVector

tol – tolerance to which the optima are computed

Output

An interval representing the range enclosure (minimum and maximum) of f over its domain dom.

Algorithm

The solver finds the global minimum of the function f over the Interval or IntervalBox dom using the Moore-Skelboe algorithm. The method actually returns an interval containing the global minimum (resp. an interval containing the global maximum), as well as a list of boxes containing the minimisers (resp. maximisers); they can be obtained calling range_and_optimisers.

This function conservatively chooses the infimum (resp. supremum) over each interval.

We refer to the documentation and source code of IntervalOptimisation for details on the implemented methods.

RangeEnclosures.enclose_TaylorModels — Function

enclose_TaylorModels(f::Function, dom::Interval_or_IntervalBox;
                     order::Int=10, normalize::Bool=true)

Compute a range enclosure using Taylor Models substitution.

Input

f – function

dom – hyperrectangular domain, either a unidimensional Interval or a multidimensional IntervalBox

order – (optional, default: 10) order of the taylor model used to compute an enclosure of f over dom

normalize – (optional, default: true) if true, normalize the taylor model on the unit symmetric box around the origin

Output

An interval representing the range enclosure (minimum and maximum) of f over its domain dom.

Algorithm

The result is obtained by an evaluation of the function over Taylor model variables after recentering in the given domain and (optionally) normalizing in the symmetric $[-1..1]^n$ domain.

We refer to the documentation and source code of TaylorModels for details.

RangeEnclosures.enclose_SumOfSquares — Function

enclose_SumOfSquares(f::Function, dom::Interval_or_IntervalBox,
                     order::Int=5, backend=SDPA.Optimizer; kwargs...)

Compute a range enclosure using sum-of-squares optimization.

Input

f – function

dom – hyperrectangular domain, either a unidimensional Interval or a multidimensional IntervalBox

order – (optional, default: 5) maximum degree of the SDP relaxation

backend – (optional, default: SDPA.Optimizer) the optimization backend (aka JuMP solver)

kwargs – additional keyword arguments

Output

An interval representing the range enclosure (minimum and maximum) of f over its domain dom.

Algorithm

The range enclosure is computed using polynomial optimization methods. We refer to the documentation and examples of SumOfSquares.jl for details.

Notes

Use the backend keyword argument to pass an SDP solver backend of your choice; additional arguments to the solver can be passed in kwargs.

For instance, to use SDPA (default choice), use it as:

julia> backend = SDPA.Optimizer

julia> enclose_SumOfSquares(f, dom, order, backend=backend)
...

To use MOSEK in non-verbose mode, let

julia> using MosekTools

julia> backend = MosekTools.Mosek.Optimizer;

julia> enclose_SumOfSquares(f, dom, order, backend=backend, QUIET=true)
...

To get the runtime, use MOI.get(model, MOI.SolveTime()).

Specifics to interval optimization

RangeEnclosures.range_and_optimisers — Function

range_and_optimisers(f::Function, dom::Interval_or_IntervalBox;
                     structure=HeapedVector, tol=1e-3)

Return an interval that contains the global minimum (resp. an interval that contains the global maximum) of a univariate or multivariate function f, as well as a list of boxes containing the minimisers (resp. minimisers).

Input

f – function

dom – hyperrectangular domain, either a unidimensional Interval or a multidimensional IntervalBox

structure – (optional, default: HeapedVector) the way in which vector elements are kept arranged; possible options are HeapedVector and SortedVector

tol – tolerance to which the optima are computed

Output

The 4-tuple (global_min, minimisers, global_max, maximisers), where:

global_min: interval that is guaranteed to contain the global minimum of f over dom

minimisers: list of boxes containing the minimisers of f over dom

global_max: interval that is guaranteed to contain the global maximum of f over dom

maximisers: list of boxes containing the maximisers of f over dom

Algorithm

The solver finds the global minimum of the function f over the Interval or IntervalBox dom using the Moore-Skelboe algorithm.

We refer to the documentation and source code of IntervalOptimisation for details.

Specifics to sum-of-squares optimization

RangeEnclosures.new_sos — Function

new_sos(backend; kwargs...)

Return a new (empty) sum-of-squares optimization problem for the given backend.

Input

backend – the backend (also called JuMP solver)

kwargs – additional keyword arguments

Output

An instance of SOSModel for the given backend and options.